drivers: power: Fix audio while charging
Instead of killing an entire feature for the sake of making audio work while charing, let's revert a spurious change.

Marvin suggested that this issue may due to a race condition. Unfortunately, i'm not experienced enough to triage this.

You can also alternatively fix this by unsetting 'CONFIG_REVERSE_CHARGE', although i'm not aware of the consequences of the same.

Ref: https://github.com/Stratosphere-Kernel/android_kernel_xiaomi_surya/commit/770deaaf8b77eabd0653ddac96657d0bab4e17a8
Signed-off-by: taalojarvi <sreedevan05@gmail.com>